+++ src/main/java/org/confng/sources/SystemPropertySource.java
@@ -11,7 +11,6 @@ import java.util.Properties;
  * 
  * @author Bharat Kumar Malviya
  * @author GitHub: github.com/imBharatMalviya
- * @version 1.0.0
  * @since 1.0
  * @see org.confng.sources.ConfigSource
  */
@@ -43,5 +42,10 @@ public class SystemPropertySource implements ConfigSource {
     @Override
     public Optional<String> get(String key) {
         return Optional.ofNullable(properties.getProperty(key));
+    }
+
+    @Override
+    public int getPriority() {
+        return 50; // Medium priority for system properties
     }
 }