+++ src/test/java/org/confng/ConfNGTest.java
@@ -21,7 +21,6 @@ import static org.testng.Assert.*;
  * 
  * @author Bharat Kumar Malviya
  * @author GitHub: github.com/imBharatMalviya
- * @version 1.0.0
  * @since 1.0
  */
 public class ConfNGTest {
@@ -149,5 +148,3 @@ public class ConfNGTest {
         System.clearProperty("browser");
     }
 }
-
-